initial: push basic module and sample file
This commit gathers the basic files required to manage/use this module.
Its purpose is to create an S3 bucket if need be, and the same for a
code repository, it can be used to create one or both, or none of those.

A sample is also provided to ease the use of this module.